+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 arena-unity notable changes. Started 2021-11-30 (version 0.0.1).
 
+## [0.14.0](https://github.com/arenaxr/arena-unity/compare/v0.13.0...v0.14.0) (2024-08-09)
+
+
+### Features
+
+* **scene-options:** load navMesh model in scene ([7e8182a](https://github.com/arenaxr/arena-unity/commit/7e8182a082ef98a9e41442a970a6ca644ac98db2))
+
+
+### Bug Fixes
+
+* **auth:** suggest creating account for 401/403 http responses ([5bc564f](https://github.com/arenaxr/arena-unity/commit/5bc564f1efea43271723c8690d5495db706f440e))
+
 ## [0.13.0](https://github.com/arenaxr/arena-unity/compare/v0.12.5...v0.13.0) (2024-06-12)